requirements.txt
setuptools
fastapi
uvicorn
faiss-cpu
numpy
psycopg2-binary
pydantic
redis
python-dotenv>=1.0.0
fakeredis>=2.23.0
face_recognition
Pillow
python-multipart
opencv-python
scikit-image
scipy